% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_json_logs}
\alias{read_json_logs}
\title{Read a directory containing JSON logs}
\usage{
read_json_logs(path)
}
\arguments{
\item{path}{Path of the directory containing JSON files.}
}
\value{
a list of \code{data.table}
}
\description{
Read a directory containing JSON logs
}
