% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_rds_logs}
\alias{read_rds_logs}
\title{Read a directory containing RDS logs}
\usage{
read_rds_logs(path)
}
\arguments{
\item{path}{Path of the directory containing RDS files or a vector of path to RDS files.}
}
\value{
a \code{list} of \code{data.table}
}
\description{
Read a directory containing RDS logs
}
\examples{
\dontrun{
# Read all RDS in a directory
logs <- read_rds_logs(path = "path/to/directory")

# Read a single file
logs <- read_rds_logs(path = "path/to/log.rds")
}
}
