% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{store_json}
\alias{store_json}
\title{Use JSON files as storage mode}
\usage{
store_json(path = "shinylogs")
}
\arguments{
\item{path}{Path where to write JSON files.}
}
\description{
Use JSON files as storage mode
}
