% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{store_rds}
\alias{store_rds}
\title{Use RDS files as storage mode}
\usage{
store_rds(path = "shinylogs")
}
\arguments{
\item{path}{Path where to write RDS files.}
}
\description{
Use RDS files as storage mode
}
