% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{store_sqlite}
\alias{store_sqlite}
\title{Use SQLite database as storage mode}
\usage{
store_sqlite(path = "shinylogs")
}
\arguments{
\item{path}{Path to the SQLite file or a directory where to create one.}
}
\description{
Use SQLite database as storage mode
}
