% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{track_usage}
\alias{track_usage}
\title{Track usage of a Shiny app}
\usage{
track_usage(storage_mode = store_json(), exclude_input_regex = NULL,
  exclude_input_id = NULL, on_unload = FALSE, exclude_users = NULL,
  get_user = NULL, session = getDefaultReactiveDomain())
}
\arguments{
\item{storage_mode}{Storage mode to use : \code{\link{store_json}}.}

\item{exclude_input_regex}{Regular expression to exclude inputs from tracking.}

\item{exclude_input_id}{Vector of \code{inputId} to exclude from tracking.}

\item{on_unload}{Logical, save log when user close the browser window or tab,
if \code{TRUE} it prevent to create \code{shinylogs}
input during normal use of the application, there will
be created only on close, downside is that a popup will appear asking to close the page.}

\item{exclude_users}{Character vectors of user for whom it is not necessary to save the log.}

\item{get_user}{A \code{function} to get user name, it should return a character and take one argument: the Shiny session.}

\item{session}{The shiny session.}
}
\description{
Used in Shiny \code{server} it will save everything that happens in a Shiny app.
}
\note{
The following \code{input}s will be accessible in the server:

  - \strong{.shinylogs_lastInput} : last \code{input} used by the user

  - \strong{.shinylogs_input} : all \code{input}s send from the browser to the server

  - \strong{.shinylogs_error} : all errors generated by \code{output}s elements

  - \strong{.shinylogs_output} : all \code{output}s generated from the server

  - \strong{.shinylogs_browserData} : information about the browser where application is displayed.
}
