% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{tracking_ui}
\alias{tracking_ui}
\title{Insert dependencies to track usage of a Shiny app}
\usage{
tracking_ui()
}
\description{
If used in \code{ui} of an application, this will create new \code{input}s available in the server.
}
\note{
The following \code{input}s will be accessible in the server:

  - \strong{.shinylogs_lastinput} : last \code{input} used by the user

  - \strong{.shinylogs_input} : all \code{input}s send from the browser to the server

  - \strong{.shinylogs_error} : all errors generated by \code{output}s elements

  - \strong{.shinylogs_output} : all \code{output}s generated from the server
}
