% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{track_usage}
\alias{track_usage}
\title{Track usage of a Shiny app}
\usage{
track_usage(storage_mode = store_json(), exclude_input_regex = NULL,
  exclude_input_id = NULL, on_unload = FALSE, exclude_users = NULL,
  session = getDefaultReactiveDomain())
}
\arguments{
\item{storage_mode}{Storage mode to use : \code{\link{store_json}}.}

\item{exclude_input_regex}{Regular expression to exclude inputs from tracking.}

\item{exclude_input_id}{Vector of \code{inputId} to exclude from tracking.}

\item{on_unload}{Logical, save log when user close the browser window or tab,
if \code{TRUE} it prevent to create \code{shinylogs}
input during normal use of the application, there will
be created only on close, downside is that a popup will appear asking to close the page.}

\item{exclude_users}{Character vectors of user for whom it is not necessary to save the log.}

\item{session}{The shiny session.}
}
\description{
Used in Shiny \code{server} it will save everything that happens in a Shiny app.
}
