% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{track_usage}
\alias{track_usage}
\title{Track usage of a Shiny app}
\usage{
track_usage(storage_mode = store_json(), exclude_users = NULL,
  session = getDefaultReactiveDomain())
}
\arguments{
\item{storage_mode}{Storage mode to use : \code{\link{store_json}}}

\item{exclude_users}{Character vectors of user for whom it is not necessary to save the log.}

\item{session}{The shiny session.}
}
\description{
Used in Shiny \code{server} it will save everything that happens in a Shiny app.
}
